drivers/snmp-ups.c: extract snmp-ups-helpers.c that can be built independently
