+++ drivers/snmp-ups.h
@@ -289,7 +289,9 @@ const char *su_find_infoval(info_lkp_t *oid2info, void *value);
 long su_find_valinfo(info_lkp_t *oid2info, const char* value);
 const char *su_find_strval(info_lkp_t *oid2info, void *value);
 
-/* Common conversion structs (functions) */
+/* Common conversion structs and functions provided by snmp-ups-helpers.c
+ * so they can be used and so "shared" by different subdrivers
+ */
 const char *su_usdate_to_isodate_info_fun(void *raw_date);
 extern info_lkp_t su_convert_to_iso_date_info[];
 /* Name the mapping location in that array for consumers to reference */
